+++ packages/isaacscript-common/src/features/playerInventory.ts
@@ -1,8 +1,4 @@
-import {
-  ActiveSlot,
-  CollectibleType,
-  ModCallback,
-} from "isaac-typescript-definitions";
+import { CollectibleType, ModCallback } from "isaac-typescript-definitions";
 import { DefaultMap } from "../classes/DefaultMap";
 import { ModUpgraded } from "../classes/ModUpgraded";
 import { ModCallbackCustom } from "../enums/ModCallbackCustom";
@@ -10,12 +6,12 @@ import { errorIfFeaturesNotInitialized } from "../featuresInitialized";
 import { arrayRemoveInPlace, copyArray } from "../functions/array";
 import { isActiveCollectible } from "../functions/collectibles";
 import { getCollectibleSet } from "../functions/collectibleSet";
+import {
+  defaultMapGetPlayer,
+  mapSetPlayer,
+} from "../functions/playerDataStructures";
 import { getAllPlayers, getPlayerIndex } from "../functions/playerIndex";
 import { repeat } from "../functions/utils";
-import {
-  PickingUpItem,
-  pickingUpItemIsCollectible,
-} from "../types/PickingUpItem";
 import { PlayerIndex } from "../types/PlayerIndex";
 import { saveDataManager } from "./saveDataManager/exports";
 
@@ -28,15 +24,6 @@ const v = {
       CollectibleType[],
       [player: EntityPlayer]
     >((player: EntityPlayer) => newPlayerInventory(player)),
-
-    /**
-     * We also keep track of the active items that the player has, so that we can easily flush them
-     * from the inventory when other active items are picked up.
-     */
-    playersActiveCollectibleTypes: new DefaultMap<
-      PlayerIndex,
-      CollectibleType[]
-    >(() => []),
   },
 };
 
@@ -54,13 +41,25 @@ function newPlayerInventory(player: EntityPlayer) {
   return inventory;
 }
 
+function resetInventory(player: EntityPlayer) {
+  const inventory = newPlayerInventory(player);
+  mapSetPlayer(v.run.playersInventory, player, inventory);
+}
+
 /** @internal */
 export function playerInventoryInit(mod: ModUpgraded): void {
   saveDataManager("playerInventory", v);
 
   mod.AddCallback(ModCallback.POST_USE_ITEM, useItemD4, CollectibleType.D4); // 3
   mod.AddCallback(ModCallback.POST_GAME_STARTED, postGameStarted); // 15
-  mod.AddCallbackCustom(ModCallbackCustom.POST_ITEM_PICKUP, postItemPickup);
+  mod.AddCallbackCustom(
+    ModCallbackCustom.POST_PLAYER_COLLECTIBLE_ADDED,
+    postCollectibleAdded,
+  );
+  mod.AddCallbackCustom(
+    ModCallbackCustom.POST_PLAYER_COLLECTIBLE_REMOVED,
+    postCollectibleRemoved,
+  );
 }
 
 // ModCallback.POST_USE_ITEM (3)
@@ -70,59 +69,37 @@ function useItemD4(
   _rng: RNG,
   player: EntityPlayer,
 ) {
-  const playerIndex = getPlayerIndex(player);
-  const inventory = newPlayerInventory(player);
-  v.run.playersInventory.set(playerIndex, inventory);
+  resetInventory(player);
 }
 
 // ModCallback.POST_GAME_STARTED (15)
 function postGameStarted() {
-  // We don't use the PostPlayerInit function because some items are not given to the player at that
+  // We don't use the PostPlayerInit callback because some items are not given to the player at that
   // point.
   for (const player of getAllPlayers()) {
     const playerIndex = getPlayerIndex(player);
     if (!v.run.playersInventory.has(playerIndex)) {
-      const inventory = newPlayerInventory(player);
-      v.run.playersInventory.set(playerIndex, inventory);
+      resetInventory(player);
     }
   }
 }
 
-// ModCallbackCustom.POST_ITEM_PICKUP
-function postItemPickup(player: EntityPlayer, pickingUpItem: PickingUpItem) {
-  if (!pickingUpItemIsCollectible(pickingUpItem)) {
-    return;
-  }
-
-  addCollectibleToInventory(player, pickingUpItem.subType);
-}
-
-function addCollectibleToInventory(
+// ModCallbackCustom.POST_PLAYER_COLLECTIBLE_ADDED
+function postCollectibleAdded(
   player: EntityPlayer,
   collectibleType: CollectibleType,
 ) {
-  const playerIndex = getPlayerIndex(player);
-  const inventory = v.run.playersInventory.getAndSetDefault(
-    playerIndex,
-    player,
-  );
+  const inventory = defaultMapGetPlayer(v.run.playersInventory, player, player);
   inventory.push(collectibleType);
+}
 
-  if (!isActiveCollectible(collectibleType)) {
-    return;
-  }
-
-  const activeCollectibleTypes =
-    v.run.playersActiveCollectibleTypes.getAndSetDefault(playerIndex);
-  activeCollectibleTypes.push(collectibleType);
-
-  const droppedActiveCollectibleTypes = activeCollectibleTypes.filter(
-    (activeCollectibleType) => !player.HasCollectible(activeCollectibleType),
-  );
-  for (const activeCollectibleType of droppedActiveCollectibleTypes) {
-    arrayRemoveInPlace(inventory, activeCollectibleType);
-    arrayRemoveInPlace(activeCollectibleTypes, activeCollectibleType);
-  }
+// ModCallbackCustom.POST_PLAYER_COLLECTIBLE_REMOVED
+function postCollectibleRemoved(
+  player: EntityPlayer,
+  collectibleType: CollectibleType,
+) {
+  const inventory = defaultMapGetPlayer(v.run.playersInventory, player, player);
+  arrayRemoveInPlace(inventory, collectibleType);
 }
 
 /**
@@ -143,11 +120,7 @@ export function getPlayerInventory(
 ): CollectibleType[] {
   errorIfFeaturesNotInitialized(FEATURE_NAME);
 
-  const playerIndex = getPlayerIndex(player);
-  const inventory = v.run.playersInventory.getAndSetDefault(
-    playerIndex,
-    player,
-  );
+  const inventory = defaultMapGetPlayer(v.run.playersInventory, player, player);
 
   const copiedInventory = copyArray(inventory);
   if (includeActiveCollectibles) {
@@ -157,31 +130,4 @@ export function getPlayerInventory(
   return copiedInventory.filter(
     (collectibleType) => !isActiveCollectible(collectibleType),
   );
-}
-
-/**
- * Helper function to add a collectible to a player. Use this instead of the
- * `EntityPlayer.AddCollectible` method if you want the collectible that is added to be
- * automatically tracked by the player inventory tracker feature.
- *
- * You only need to use this function if you are using the inventory feature from the standard
- * library.
- */
-export function addCollectible(
-  player: EntityPlayer,
-  collectibleType: CollectibleType,
-  charge?: int,
-  firstTimePickingUp?: boolean,
-  activeSlot?: ActiveSlot,
-  varData?: int,
-): void {
-  player.AddCollectible(
-    collectibleType,
-    charge,
-    firstTimePickingUp,
-    activeSlot,
-    varData,
-  );
-
-  addCollectibleToInventory(player, collectibleType);
 }