+++ packages/isaacscript-common/src/functions/array.ts
@@ -45,7 +45,7 @@ export function arrayRemove<T>(
 
 /**
  * Removes the specified element(s) from the array. If the specified element(s) are not found in the
- * array, this function will do nothing. Returns whether or not one or more elements were removed.
+ * array, this function will do nothing. Returns true if one or more elements were removed.
  *
  * This function is variadic, meaning that you can specify N arguments to remove N elements.
  */
@@ -90,7 +90,7 @@ export function arrayRemoveIndex<T>(
 
 /**
  * Removes the elements at the specified indexes from the array. If the specified indexes are not
- * found in the array, this function will do nothing. Returns whether or not one or more elements
+ * found in the array, this function will do nothing. Returns true if one or more elements
  * were removed.
  *
  * This function is variadic, meaning that you can specify N arguments to remove N elements.